Fix Jedis issue and add tests
Jedis instances were not released which caused the program to block at some point.
Added tests for the simulators (not done yet).

+++ src/main/java/company/product/Product.java
@@ -14,7 +14,7 @@ public class Product extends JedisObject {
     private GeoCoordinate currentLocation;
     private float price;
 
-    public Product(ProductType productType, GeoCoordinate currentLocation, int price) {
+    public Product(ProductType productType, GeoCoordinate currentLocation, float price) {
         this.currentLocation = currentLocation;
         this.productType = productType;
         this.price = price;
